+++ php/src/UAParser/Result/Device.php
@@ -1,0 +1,19 @@
+<?php
+/**
+ * ua-parser
+ *
+ * Copyright (c) 2011-2012 Dave Olsen, http://dmolsen.com
+ *
+ * Released under the MIT license
+ */
+namespace UAParser\Result;
+
+class Device extends AbstractSoftware
+{
+    /** @var string */
+    public $brand;
+
+    /** @var string */
+    public $model;
+    
+}